% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KWA1Dalt.R
\name{KWA1D.estimator}
\alias{KWA1D.estimator}
\title{Adaptive Univariate Kernel Weighted Average Estimation}
\usage{
KWA1D.estimator(x, df.search)
}
\arguments{
\item{x}{numeric vector containing the dataset.}

\item{df.search}{numeric vector representing a grid to search for ``best" \code{df} match.}
}
\value{
A list consisting of KWA location and scale estimates together with the ``best" \code{df} and associated maximal
\code{p-value} from the Cramer-von Mises goodness-of-fit test.
}
\description{
Computes Kernel Weighted Average estimates of univariate Student's \emph{t} location and scale
}
\details{
See \code{\link{KWA1D}}
}
\note{
If \code{df.search} is not provided, the default search interval is
\code{\strong{c(seq(from = 1.0, to = 5,  by = 0.2), seq(from = 5.5, to = 15, by = 0.5), seq(from = 16,  to = 30, by = 1),
50, 100, 500, Inf)}}
}
\examples{
n <- 100
df <- 2
x <- rt(n, df)
KWA1D.estimator(x) #using default search interval
KWA1D.estimator(x, df.search = seq(1, 5, by = 0.05)) #using a user-provided search interval

}
\references{
Pokojovy, M., Chen, S., Anum, A.T and Koomson, J. (2023) Adaptive location and scale estimation with kernel weighted averages. (submitted)

Rousseeuw, P.J. and Croux, C. (1993) Alternatives to the Median Absolute Deviation. \emph{Journal of the
American Statistical Association} \strong{88}, 1273–1283.

Rousseeuw, P. J. and Leroy, A. M. (1987) \emph{Robust Regression and Outlier Detection}. Wiley.
}
\seealso{
\code{\link{KWA1D}}, \code{\link[robustbase]{Qn}}
}
\author{
Michael Pokojovy (\email{michael.pokojovy@gmail.com}), Su Chen, Andrews T. Anum, John Koomson
}
\keyword{KWA}
\keyword{kernel weighted average}
\keyword{Student t-distribution}
\keyword{bandwidth selection}
\keyword{density functional}
\keyword{heavy tails}
