% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KWA1D.R
\name{KWA1D}
\alias{KWA1D}
\title{Univariate Kernel Weighted Average Estimation}
\usage{
KWA1D.loc(x, df, bw = NULL, adjust = 1.0)
KWA1D.scale(x, df, bw.loc = NULL, adjust.loc = 1.0, bw.scale = NULL, adjust.scale = 1.0, bias.corr = NULL)
KWA1D.loc.scale(x, df, bw.loc = NULL, adjust.loc = 1.0, bw.scale = NULL, adjust.scale = 1.0, bias.corr = NULL)
}
\arguments{
\item{x}{numeric vector containing the dataset.}

\item{df}{degrees of freedom.}

\item{bw, bw.loc}{bandwidth for location estimation.}

\item{bw.scale}{bandwidth for scale estimation.}

\item{adjust, adjust.loc}{scalar adjustment to location bandwidth; default value is 1.0.}

\item{adjust.scale}{scalar adjustment to scale bandwidth; default value is 1.0.}

\item{bias.corr}{scalar factor to make the scale estimate unbiased.}
}
\value{
A scalar representing KWA location or scale estimate. When \code{KWA1D.loc.scale} is called, a list containing
location and scale estimates is returned.
}
\description{
Computes location and scale estimates for a univariate Student's t dataset via Kernel Weighted Average estimation
}
\details{
Consider a location-scale family \eqn{f(x|\mu, \sigma) = 1/\sigma \times f_0((x-\mu)/\sigma)} for real \eqn{x},
where \eqn{\mu} and \eqn{\sigma} are the location and scale parameters, respectively, and \eqn{f_0(\cdot)} is symmetric about 0.
For a squared-integrable random variable \eqn{X}, the location and scale parameters can be expressed as
\deqn{\mu = (\int x f^2(x) dx) /  (\int f^2(x) dx), }
\deqn{\sigma = ( ( \int z^2 f^2_0(z) dz) /  ( \int f^2_0 (z) dz) )^{-1/2} \times  ( ( \int (x-\mu)^2 f^2(x) dx) /  ( \int f^2(x) dx) )^{1/2}.}
These functionals can be estimated using appropriate KDE-based U-statistics leading to KWA estimates for location and scale parameters.
}
\note{
If \code{bw} is not provided, \code{KWA1D.loc} finds an optimal bandwidth based on \code{df} and the sample size.
Similarly, \code{KWA1D.scale} and \code{KWA1D.loc.scale} compute \code{bw.loc} and \code{bw.scale} if they are not provided.
}
\examples{
n <- 100
df <- 2
x <- rt(n, df)
loc <- KWA1D.loc(x, df)
scale <- KWA1D.scale(x, df)
KWA1D.loc.scale(x, df)

}
\references{
Pokojovy, M., Chen, S., Anum, A.T and Koomson, J. (2024) Adaptive location and scale estimation with kernel weighted averages. \emph{Communications in Statistics: Simulation and Computation} (accepted)

Rousseeuw, P.J. and Croux, C. (1993) Alternatives to the median absolute deviation. \emph{Journal of the
American Statistical Association} \strong{88}, 1273–1283.

Rousseeuw, P. J. and Leroy, A. M. (1987) \emph{Robust Regression and Outlier Detection}. Wiley.
}
\seealso{
\code{\link{KWA1D.estimator}}, \code{\link[robustbase]{Qn}}
}
\author{
Michael Pokojovy (\email{michael.pokojovy@gmail.com}), Su Chen, Andrews T. Anum, John Koomson
}
\keyword{KWA}
\keyword{kernel weighted average}
\keyword{Student t-distribution}
\keyword{bandwidth selection}
\keyword{density functional}
\keyword{heavy tails}
